+++ src/rxAge/rxAge.js
@@ -3,7 +3,7 @@ angular.module('encore.ui.rxAge', [])
 // age.js -- https://github.rackspace.com/gist/roge5001/2417941
 // age.spec.js -- https://github.rackspace.com/gist/roge5001/2417940
 .filter('rxAge', function () {
-    return function (dateString, maxUnits) {
+    return function (dateString, maxUnits, verbose) {
         if (!dateString) {
             return 'Unavailable';
         } else if (dateString === 'z') {
@@ -19,19 +19,38 @@ angular.module('encore.ui.rxAge', [])
         var mins = parseInt(duration.asMinutes(), 10);
         var age = [];
 
-        if (!_.isNumber(maxUnits)) {
-            maxUnits = 2;
+        if (_.isBoolean(maxUnits)) {
+            // if maxUnits is a boolean, then we assume it's meant to be the verbose setting
+            verbose = maxUnits;
+        } else if (!_.isBoolean(verbose)) {
+            // otherwise, if verbose isn't set, default to false
+            verbose =  false;
+        }
+
+        // This initialization has to happen AFTER verbose init so that we can
+        // use the original passed in value.
+        maxUnits = (_.isNumber(maxUnits)) ? maxUnits : 2;
+
+        var dateUnits = [days, hours - (24 * days), mins - (60 * hours)];
+        var suffixes = ['d', 'h', 'm'];
+
+        if (verbose) {
+            suffixes = [' day', ' hour', ' minute'];
+
+            _.forEach(suffixes, function (suffix, index) {
+                suffixes[index] += ((dateUnits[index] != 1) ? 's' : '');
+            });
         }
 
         if (days > 0) {
-            age.push({ value: days, suffix: 'd' });
+            age.push({ value: days, suffix: suffixes[0] });
         }
 
         if (hours > 0) {
-            age.push({ value: hours - (24 * days), suffix: 'h' });
+            age.push({ value: hours - (24 * days), suffix: suffixes[1] });
         }
 
-        age.push({ value: mins - (60 * hours), suffix: 'm' });
+        age.push({ value: mins - (60 * hours), suffix: suffixes[2] });
 
         return _.map(age.slice(0, maxUnits), function (dateUnit, index, listOfAges) {
             if (index == listOfAges.length - 1) {
@@ -39,6 +58,6 @@ angular.module('encore.ui.rxAge', [])
             } else {
                 return Math.floor(dateUnit.value) + dateUnit.suffix;
             }
-        }).join(' ');
+        }).join((verbose) ? ', ' : ' ');
     };
 });