+++ CHANGELOG.md
@@ -1,3 +1,35 @@
+<a name="0.9.1"></a>
+### 0.9.1 (2014-05-12)
+
+
+#### Features
+
+* **rxAge:** Added the ability to adjust verbosity ([3264c877](git@github.com:rackerlabs/encore-ui/commit/3264c877e041d181c43f2d34d5e69f2260166b14))
+
+
+<a name="0.9.0"></a>
+## 0.9.0 (2014-05-08)
+
+
+#### Bug Fixes
+
+* **rxNotify:** change error icon to avoid UX issue ([11facc8f](git@github.com:rackerlabs/encore-ui/commit/11facc8f804b92f0752abf3dcd8a89728d31b8fa))
+* **rxUnauthorizedInterceptor:** Redirect back to originator ([61fe1a30](git@github.com:rackerlabs/encore-ui/commit/61fe1a30c29e967b30444f56ac8422c01d568b3a))
+
+
+#### Features
+
+* **rxApp:**
+  * switch rxAppRoutes to service ([1c29cf73](git@github.com:rackerlabs/encore-ui/commit/1c29cf735f63eb1880d457cf1d1861a41ef7a35a))
+  * expose appRoutes via rootScope ([800f225b](git@github.com:rackerlabs/encore-ui/commit/800f225b74bf479bfd68442a794b1aaf016c3e6f))
+
+
+#### Breaking Changes
+
+* Creation and editing of menus for rxApp has undergone major changes. See [PR #223](https://github.com/rackerlabs/encore-ui/pull/233) for details
+ ([800f225b](git@github.com:rackerlabs/encore-ui/commit/800f225b74bf479bfd68442a794b1aaf016c3e6f))
+
+
 <a name="0.8.4"></a>
 ### 0.8.4 (2014-05-07)